possible to compile code now

+++ backend/workoutapp/src/main/java/org/example/workoutapp/model/ActivityWorkoutExercise.java
@@ -18,7 +18,7 @@ public class ActivityWorkoutExercise {
 
     private int exerciseReps;
 
-    private double exerciseWeight;
+    private int exerciseWeight;
 
     @ManyToOne
     @JoinColumn(name = "activityId")
